The port is a u16 where the kernel is
`listen_port` is `u16` in the netlink types and the struct that carries
it out of `observed` said `u32`. Nothing on this laptop compiles that
function — it is `#[cfg(target_os = "linux")]` — so all three of fmt,
clippy and test passed and the cross-build for the node did not.

The gate is only the gate for the code the compiler in front of it can
see.

+++ src/network/tunnel.rs
@@ -491,7 +491,7 @@ pub fn observed() -> Result<Interface, String> {
 /// answer nobody needs.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Interface {
-    pub port: u32,
+    pub port: u16,
     pub peers: Vec<Handshake>,
 }
